requirements.txt
elasticsearch~=8.7.0
python-dotenv~=1.0.0
psutil==5.9.5